+++ auth-manage-service/src/main/java/com/byd/auth/manage/service/controller/SysBaseController.java
@@ -84,7 +84,7 @@ public class SysBaseController {
             defaultValue = ""),
         @ApiImplicitParam(name = "remark", value = "备注信息", dataType = DataType.STRING, paramType = ParamType.BODY,
             defaultValue = ""),
-        @ApiImplicitParam(name = "delete_flag", value = "逻辑删除", dataType = DataType.BYTE, paramType = ParamType.BODY,
+        @ApiImplicitParam(name = "delete_flag", value = "逻辑删除", dataType = DataType.INT, paramType = ParamType.BODY,
             defaultValue = "0")})
     public Object insertOrUpdateSysBase(@RequestBody SysBaseVo sysBaseVo) {
         return sysBaseService.insertOrUpdateSysBase(sysBaseVo);
